label management refactoring for reusable component
