+++ communitter-frontend/src/utilities/config.js
@@ -2,6 +2,6 @@ export const url = "http://localhost:8080";
 export const tokenName = "CommunitterToken";
 export const defaultFetchOpts = {
   method: "GET",
-  headers: {},
+  headers:{}
 };
 export const wikidataUrl = "https://www.wikidata.org/w/api.php?";
