+++ communitter-frontend/src/utilities/fetchWithOptions.js
@@ -3,6 +3,8 @@ import { tokenName } from "./config";
 export async function fetchWithOpts(url, options) {
   const token = localStorage.getItem(tokenName);
   options.headers.Authorization = "Bearer " + token;
+  
+  
   const response = await fetch(url, options);
   if (!response.ok) {
     const message = await response.text();
